test split emulator e2e contracts
