+++ tests/e2e/utils_test.go
@@ -21,8 +21,6 @@ import (
 	"github.com/btcsuite/btcd/wire"
 )
 
-const counterPacketType = 2
-
 type artifact struct {
 	Name      string          `json:"contractName"`
 	Functions []functionGroup `json:"functions"`
@@ -65,54 +63,6 @@ func (f *testPrevOutFetcher) FetchVtxoPrevOutPkScript(outpoint wire.OutPoint) []
 		return nil
 	}
 	return tx.TxOut[outpoint.Index].PkScript
-}
-
-func TestCompiledHTLCCovenant(t *testing.T) {
-	contract := compileArtifact(t, filepath.Join("..", "..", "examples", "htlc", "htlc.ark"))
-	serverKey := fixedPublicKey(1)
-	emulatorKey := fixedPublicKey(2)
-	claim := instantiateGroup(t, contract, "claim", map[string][]byte{
-		"preimageHash": bytes.Repeat([]byte{0x42}, 20),
-	}, serverKey, emulatorKey)
-
-	prevTx := fundingTx(claim.pkScript, 10_000)
-
-	valid := spendingPSBT(t, prevTx, claim, 10_000, claim.pkScript)
-	requireVMResult(t, valid, emulatorKey, true)
-
-	underfunded := spendingPSBT(t, prevTx, claim, 9_999, claim.pkScript)
-	requireVMResult(t, underfunded, emulatorKey, false)
-}
-
-func TestCompiledCounterRecursion(t *testing.T) {
-	contract := compileArtifact(t, filepath.Join("testdata", "counter.ark"))
-	serverKey := fixedPublicKey(1)
-	emulatorKey := fixedPublicKey(2)
-	counter := instantiateGroup(t, contract, "increment", nil, serverKey, emulatorKey)
-
-	deployment := fundingTx(counter.pkScript, 20_000)
-	addCounterPacket(t, deployment, 0)
-
-	first := spendingPSBT(t, deployment, counter, 20_000, counter.pkScript, counterPacket(t, 1))
-	requireVMResult(t, first, emulatorKey, true)
-
-	skippedState := spendingPSBT(t, deployment, counter, 20_000, counter.pkScript, counterPacket(t, 2))
-	requireVMResult(t, skippedState, emulatorKey, false)
-
-	second := spendingPSBT(
-		t, first.UnsignedTx, counter, 20_000, counter.pkScript, counterPacket(t, 2),
-	)
-	requireVMResult(t, second, emulatorKey, true)
-
-	wrongScript := spendingPSBT(
-		t, first.UnsignedTx, counter, 20_000, p2trScript(t, []byte{txscript.OP_TRUE}), counterPacket(t, 2),
-	)
-	requireVMResult(t, wrongScript, emulatorKey, false)
-
-	underfunded := spendingPSBT(
-		t, first.UnsignedTx, counter, 19_999, counter.pkScript, counterPacket(t, 2),
-	)
-	requireVMResult(t, underfunded, emulatorKey, false)
 }
 
 func compileArtifact(t *testing.T, source string) artifact {
@@ -362,26 +312,6 @@ func executeArkadeScripts(ptx *psbt.Packet, emulatorKey *btcec.PublicKey) error 
 		}
 	}
 	return nil
-}
-
-func addCounterPacket(t *testing.T, tx *wire.MsgTx, value uint64) {
-	t.Helper()
-	packet := counterPacket(t, value)
-	ext := extension.Extension{packet}
-	output, err := ext.TxOut()
-	if err != nil {
-		t.Fatalf("counter packet: %v", err)
-	}
-	tx.AddTxOut(output)
-}
-
-func counterPacket(t *testing.T, value uint64) extension.Packet {
-	t.Helper()
-	payload, err := arkade.BigNumFromUint64(value + 1).Bytes()
-	if err != nil {
-		t.Fatalf("counter payload: %v", err)
-	}
-	return extension.UnknownPacket{PacketType: counterPacketType, Data: payload}
 }
 
 func fixedPublicKey(value byte) *btcec.PublicKey {
